requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_codedeploy
aws_cdk.aws_lambda 
aws_cdk.aws_codebuild 
aws_cdk.aws_codepipeline
aws_cdk.aws_codecommit 
aws_cdk.aws_codepipeline_actions 
aws_cdk.aws_s3
